+++ CHANGELOG.md
@@ -13,6 +13,10 @@
 * Parameters are converted to floats, unwrapping interface data types.
   [(#41)](https://github.com/PennyLaneAI/PennyLane-IonQ/pull/41)
 
+* If response returns as failure, an error is raised. If the user
+  submits an empty circuit, a warning is raised.
+  [(#43)](https://github.com/PennyLaneAI/PennyLane-IonQ/pull/43)
+
 ### Contributors
 
 This release contains contributions from (in alphabetical order):